takes vec values for joystick as floats.

+++ src/device.c
@@ -47,7 +47,7 @@ void write_key_event(int code, int value){
 	libevdev_uinput_write_event(uidev, EV_KEY, code, value);
 }
 
-void write_abs_event(int code, int value){
+void write_abs_event(int code, float value){
 	libevdev_uinput_write_event(uidev, EV_ABS, code, value);
 }
 void sync_events(){
@@ -69,7 +69,7 @@ void click( int code){
 	release(code);
 }
 
-void move_abs_event( int code, int pos){
+void move_abs_event( int code, float pos){
 	write_abs_event(code, pos);
 }
